Fixed error if we use Postgres database
As Laravel documantation mentioned for DB class method insertGetId:

When using PostgreSQL the insertGetId method expects the auto-incrementing column to be named id. If you would like to retrieve the ID from a different "sequence", you may pass the column name as the second parameter to the insertGetId method.

In context of code, returned ID's isn't used, we don't need it.